requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
arrow==1.1.0
asgiref==3.3.1
async-generator==1.10
attr==0.3.1
attrs==20.3.0
backcall==0.2.0
bleach==3.2.1
camelcase==0.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
clr==1.0.3
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.5
django-api-gateway==0.0.7
django-cors-headers==3.7.0
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
django-ok-cart==0.0.2
django-rest-knox==4.1.0
django-shopping-cart==0.1
djangorestframework==3.12.4
docx-utils==0.1.3
entrypoints==0.3
et-xmlfile==1.1.0
fastapi==0.65.1
Flask==2.0.0
greenlet==1.1.0
gunicorn==20.1.0
h11==0.12.0
idna==2.10
importlib-metadata==3.3.0
ipykernel==5.4.2
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.0
jedi==0.18.0
Jinja2==3.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lxml==4.6.3
MarkupPy==1.14
MarkupSafe==2.0.0
matplotlib==3.4.2
mistune==0.8.4
mock==4.0.3
mysqlclient==2.0.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.6
numpy==1.20.0
odfpy==1.4.1
opencv-python==4.5.2.54
openpyxl==3.0.7
packaging==20.8
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.9
psycopg2==2.8.6
pycparser==2.20
pydantic==1.8.2
Pygments==2.7.3
PyMySQL==1.0.1
pyparsing==2.4.7
pypiwin32==223
pyrsistent==0.17.3
python-dateutil==2.8.1
python-docx==0.8.10
python-pptx==0.6.18
pytz==2020.5
pywin32==300
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
razorpay==1.2.0
requests==2.25.1
Send2Trash==1.5.0
six==1.15.0
SQLAlchemy==1.4.15
sqlparse==0.4.1
starlette==0.14.2
stripe==2.58.0
tablib==3.0.0
terminado==0.9.2
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
waitress==2.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
xlrd==2.0.1
XlsxWriter==1.4.0
xlwt==1.3.0
zipp==3.4.0
